+++ packages/api-client/src/components/TopNav/TopNav.vue
@@ -1,17 +1,83 @@
 <script setup lang="ts">
-import { useTopNav } from '@/store/topNav'
-import { ScalarIcon } from '@scalar/components'
-import { computed } from 'vue'
+import { ROUTES } from '@/constants'
+import { useWorkspace } from '@/store'
+import { type Icon, ScalarIcon } from '@scalar/components'
+import { capitalize } from '@scalar/oas-utils/helpers'
+import { computed, reactive, ref, watch } from 'vue'
 
 import TopNavItem from './TopNavItem.vue'
 
-const {
-  topNavItems,
-  addNavItem,
-  activeNavItemIdx,
-  setNavItemIdx,
-  removeNavItem,
-} = useTopNav()
+const { activeRequest, router } = useWorkspace()
+
+/** Nav Items list */
+const topNavItems = reactive([{ label: '', path: '', icon: 'Add' as Icon }])
+const activeNavItemIdx = ref(0)
+
+/**
+ * Logic to handle adding a nav item
+ * based on the current route
+ */
+function handleNavLabelAdd() {
+  const activeRoute = ROUTES.find((route) => {
+    return router.currentRoute.value.name == route.name
+  })
+
+  if (!activeRoute) return
+
+  // if it's a request we can push in a request
+  if (activeRoute?.name === 'request') {
+    topNavItems[activeNavItemIdx.value] = {
+      label: activeRequest.value?.summary || '',
+      path: router.currentRoute.value.path,
+      icon: activeRoute.icon,
+    }
+  } else {
+    // not requests so its the other nav items
+    // we can eventually be more granular
+    topNavItems[activeNavItemIdx.value] = {
+      label: capitalize(activeRoute?.name) || '',
+      path: router.currentRoute.value.path,
+      icon: activeRoute.icon,
+    }
+  }
+}
+
+function handleNavRoute() {
+  router.push(topNavItems[activeNavItemIdx.value].path)
+}
+
+/**
+ * adding a nav item sets the new index and nav item
+ * based on the route
+ */
+function addNavItem() {
+  topNavItems.push({ label: '', path: '', icon: 'Add' })
+  activeNavItemIdx.value = topNavItems.length - 1
+  handleNavLabelAdd()
+}
+
+function setNavItemIdx(idx: number) {
+  activeNavItemIdx.value = idx
+  handleNavRoute()
+}
+
+// when the route changes we need update the active nav item
+watch(
+  () => router.currentRoute.value.path,
+  () => {
+    handleNavLabelAdd()
+  },
+  { immediate: true },
+)
+
+function removeNavItem(idx: number) {
+  topNavItems.splice(idx, 1)
+  activeNavItemIdx.value = Math.min(
+    activeNavItemIdx.value,
+    topNavItems.length - 1,
+  )
+  handleNavRoute()
+}
 
 const activeNavItemIdxValue = computed(() => activeNavItemIdx.value)
 </script>
